% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aSimpleFile.R
\name{aSimpleFunction}
\alias{aSimpleFunction}
\title{Add together two numbers.}
\usage{
aSimpleFunction(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
This is truly a great and much-needed function
}
\examples{
aGreatFunction(1, 1)
aGreatFunction(10, 1)
}
